web/onos-gui/src/app/onos.component.spec.ts
/*
 * SPDX-FileCopyrightText: 2020-present Open Networking Foundation <info@opennetworking.org>
 *
 * SPDX-License-Identifier: Apache-2.0
 */

import { TestBed, waitForAsync } from '@angular/core/testing';
import {RouterTestingModule} from '@angular/router/testing';
import {OnosComponent} from './onos.component';

describe('OnosComponent', () => {
    beforeEach(waitForAsync(() => {
        TestBed.configureTestingModule({
            imports: [
                RouterTestingModule
            ],
            declarations: [
                OnosComponent
            ],
        }).compileComponents();
    }));

    it('should create the app', () => {
        const fixture = TestBed.createComponent(OnosComponent);
        const app = fixture.debugElement.componentInstance;
        expect(app).toBeTruthy();
    });

    it(`should have as title 'onos-gui'`, () => {
        const fixture = TestBed.createComponent(OnosComponent);
        const app = fixture.debugElement.componentInstance;
        expect(app.title).toEqual('onos-gui');
    });

    it('should render title in a h1 tag', () => {
        const fixture = TestBed.createComponent(OnosComponent);
        fixture.detectChanges();
        const compiled = fixture.debugElement.nativeElement;
        expect(compiled.querySelector('h1').textContent).toContain('Welcome to onos-gui!');
    });
});
